// src/app/pages/transaction/model/report.ts
export interface ReportModel {
  applicantAccountNo: string;
  applicantFullName: string;
  applicantAddress: string;
  beneficiaryAccountNo: string;
  formA: string;
  fxAmount: string;
  purchaseType: string;
  transactionType: string;
  purpose: string;
  taxCertNo: string;
  statusText: string;
  isCustomer: string;
  isCashTranx: string;
  createdOn: string;
}